Use Variable struct in initial Grammar representation
