+++ project.gyp
@@ -25,6 +25,7 @@
         'src/compiler/build_tables/rule_can_be_blank.cc',
         'src/compiler/compile.cc',
         'src/compiler/generate_code/c_code.cc',
+        'src/compiler/grammar.cc',
         'src/compiler/lex_table.cc',
         'src/compiler/parse_grammar.cc',
         'src/compiler/parse_table.cc',
@@ -42,7 +43,6 @@
         'src/compiler/prepare_grammar/token_description.cc',
         'src/compiler/rule.cc',
         'src/compiler/syntax_grammar.cc',
-        'src/compiler/variable.cc',
         'src/compiler/rules/blank.cc',
         'src/compiler/rules/built_in_symbols.cc',
         'src/compiler/rules/character_range.cc',
