+++ src/compiler/prepare_grammar/initial_syntax_grammar.h
@@ -6,7 +6,7 @@
 #include <set>
 #include "tree_sitter/compiler.h"
 #include "compiler/rules/symbol.h"
-#include "compiler/variable.h"
+#include "compiler/grammar.h"
 #include "compiler/syntax_grammar.h"
 
 namespace tree_sitter {
